+++ apps/client/src/components/leo/modals/manage-record/manage-record-modal.tsx
@@ -11,9 +11,9 @@ import useFetch from "lib/useFetch";
 import { ModalIds } from "types/ModalIds";
 import { useTranslations } from "use-intl";
 import { RecordType, type PenalCode, type Record, PaymentStatus } from "@snailycad/types";
-import { PenalCodesTable } from "./ManageRecord/PenalCodesTable";
-import { SelectPenalCode } from "./ManageRecord/select-penal-code";
-import { SeizedItemsTable } from "./ManageRecord/seized-items/SeizedItemsTable";
+import { PenalCodesTable } from "./penal-codes-table";
+import { SelectPenalCode } from "./select-penal-code";
+import { SeizedItemsTable } from "./seized-items/seized-items-table";
 import { toastMessage } from "lib/toastMessage";
 import { useFeatureEnabled } from "hooks/useFeatureEnabled";
 import type { PostRecordsData, PutRecordsByIdData } from "@snailycad/types/api";
