+++ apps/client/src/components/leo/modals/manage-record/penal-codes-table.tsx
@@ -2,7 +2,7 @@ import { Table, useTableState } from "components/shared/Table";
 import { useFormikContext } from "formik";
 import { useTranslations } from "next-intl";
 import type { PenalCode } from "@snailycad/types";
-import { TableItemForm } from "./TableItemForm";
+import { TableItemForm } from "./table-item-form";
 import { useFeatureEnabled } from "hooks/useFeatureEnabled";
 
 interface Props {
