+++ apps/client/src/components/leo/modals/manage-record/seized-items/manage-seized-items-modal.tsx
