+++ apps/client/src/components/leo/modals/manage-record/seized-items/seized-items-table.tsx
@@ -6,7 +6,7 @@ import { useFormikContext } from "formik";
 import { yesOrNoText } from "lib/utils";
 import { useTranslations } from "next-intl";
 import { ModalIds } from "types/ModalIds";
-import { ManageSeizedItemsModal } from "./ManageSeizedItemsModal";
+import { ManageSeizedItemsModal } from "./manage-seized-items-modal";
 import type { SeizedItem } from "@snailycad/types";
 import { useTemporaryItem } from "hooks/shared/useTemporaryItem";
