+++ apps/client/src/components/leo/modals/manage-record/table-item-form.tsx
